+++ README.md
@@ -1,3 +1,3 @@
 # Project3-LMS
 
-Lucy, Michael, and Samuel have created their own CRM to help focus Michael on his work. Aimed at small businesses, our CRM is focused on delivering services to poeples' houses. We will promote our clients' companies growth by helping them keep track of their customer's contact info (address, phone number, and email), time availability, schedule services, and tracks services provided.
+Lucy, Michael, and Samuel have created their own CRM to help focus Michael on his work. Aimed at small businesses, our CRM is focused on delivering services to peoples' houses. We will promote our clients' companies growth by helping them keep track of their customer's contact info (address, phone number, and email), time availability, schedule services, and tracks services provided.
