+++ client/src/components/Form/clientApp.js
@@ -14,9 +14,9 @@ import LeftNav from 'material-ui/lib/left-nav';
 
 injectTapEventPlugin();
 
-import ClientList from '../ClientList'
+import ClientList from './clientList'
 import InputForm from '../containers/InputForm'
-import TaskForm from '../containers/TaskForm'
+import TaskForm from './taskForm'
 import TaskList from './TaskList'
 
 export default class App extends Component {
