+++ client/src/components/Form/tasks.js
@@ -3,9 +3,9 @@ import { Component } from 'react';
 import { connect } from 'react-redux'
 import { Link } from 'react-router';
 
-
+import TaskForm from './taskForm'
 import TaskList from './TaskList'
-import TaskForm from '../containers/TaskForm'
+// import TaskForm from '../containers/TaskForm'
 
 export default class Tasks extends Component {
   render() {
